langgraph-server/pyproject.toml
[project]
name = "langgraph-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arize-phoenix-otel>=0.9.2",
    "duckduckgo-search>=8.0.1",
    "langchain-community>=0.3.21",
    "langchain-openai>=0.3.14",
    "langgraph>=0.3.31",
    "openinference-instrumentation-langchain>=0.1.41",
    "pydantic-ai>=0.1.3",
    "wikipedia>=1.4.0",
]
